requirements.txt
pandas==2.2.3
numpy==1.24.3
numpy-financial==1.0.0
plotly==6.0.1
python_dateutil==2.9.0.post0
streamlit==1.44.1
openpyxl==3.1.2